# docs/swagger.yaml
basePath: /
...
host: localhost:9001
info:
  contact: {}
  title: Backend Golang Tickitz App
  version: "1.0"
paths:
  /admin/movies:
    get:
      description: Semua data Movie untuk admin
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get All Movies (Admin)
      tags:
      - Admin-Movies
  /admin/movies/{movieId}:
    patch:
      consumes:
      - multipart/form-data
      description: Update data film berdasarkan ID dengan upload gambar opsional
      parameters:
      - description: Movie ID
        in: path
        name: movieId
        required: true
        type: integer
      - description: Judul film
        in: formData
        name: title
        type: string
      - description: Sinopsis film
        in: formData
        name: synopsis
        type: string
      - description: Durasi film dalam menit
        in: formData
        name: duration_minutes
        type: integer
      - description: Tanggal rilis (YYYY-MM-DD)
        in: formData
        name: release_date
        type: string
      - description: ID direktur
        in: formData
        name: directors_id
        type: integer
      - description: Rating film
        in: formData
        name: rating
        type: number
      - description: File gambar poster (opsional)
        in: formData
        name: poster_image
        type: file
      - description: File gambar background (opsional)
        in: formData
        name: bg_path
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update Movie (Admin)
      tags:
      - Admin-Movies
  /admin/movies/add:
    post:
      consumes:
      - multipart/form-data
      description: Tambah data movie baru dengan upload gambar
      parameters:
      - description: Judul film
        in: formData
        name: title
        required: true
        type: string
      - description: Sinopsis film
        in: formData
        name: synopsis
        required: true
        type: string
      - description: Durasi film dalam menit
        in: formData
        name: duration_minutes
        required: true
        type: integer
      - description: Tanggal rilis (YYYY-MM-DD)
        in: formData
        name: release_date
        required: true
        type: string
      - description: ID direktur
        in: formData
        name: directors_id
        required: true
        type: integer
      - description: Rating film
        in: formData
        name: rating
        type: number
      - description: File gambar poster
        in: formData
        name: poster_image
        required: true
        type: file
      - description: File gambar background
        in: formData
        name: bg_path
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Tambah Movie (Admin)
      tags:
      - Admin-Movies
  /admin/movies/delete/{movieId}:
    delete:
      description: Tandai movie sebagai deleted (soft delete) dan isi deleted_at timestamp
      parameters:
      - description: Movie ID
        in: path
        name: movieId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Soft delete movie (Admin)
      tags:
      - Admin-Movies
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"